doc: Add more information to the document, make the checklist shorter
adds the f/, h/, d/ and r/ branch conventions, describes how to use the
draft label, mentions assigning yourself to issues and PRs, linking
issues to PRs, and where in the commit the issue should be mentioned.
Also specifies that the code needs to be passed through clang-format
before merge.

Changes the checklist slightly so it only includes the final things that
need to be verified for the merge itself, and doesn't get too huge.
Other info about the PR process is instead described in a section above
the checklist.